Add files via upload
My approach involved scrutinizing the trends by examining the percentage of pediatric populations undergoing screening during hospital visits, with a particular emphasis on the subset requiring behavioral health interventions. The comprehensive examination provides valuable insights into the landscape of pediatric behavioral health.  

For this analysis, I analyzed the Mass CBHI behavioral health screening data to understand the behavioral health trends for the years 2012 to 2021. 

The Western, Central, and Boston regions in Massachusetts exhibit the most substantial rises in the percentage of screenings identifying behavioral health needs, with increases of 12.71%, 4.8%, and 5.45%, respectively. Throughout the measurement period, the Central Region in Massachusetts witnessed a noteworthy decrease of 4.11% in the screening rate for behavioral health, specifically measured as the percentage of visits incorporating a behavioral health screen.  

To gain a comprehensive insight into the prevailing patterns of behavioral health conditions such as Attention-Deficit/Hyperactivity Disorder (ADHD), autism spectrum disorder (ASD), and incidents of bullying, at the state level in Massachusetts and on a national scale. Pertaining to ADHD and ASD, the focus was directed towards assessing the proportion of the population currently afflicted by these conditions. In the case of bullying, the analysis considered the percentage of children subjected to bullying incidents occurring 1-2 times per month. The analysis is below. The states of Louisiana, West Virginia, and New Hampshire fall under the top 10 for the prevalence of both ADHD and autism spectrum disorder ASD.  Amongst these, Louisiana has the highest proportion of ADHD prevalence with 15.7% ADHD issues reported at the current time.   
Massachusetts is positioned 3rd among states in terms of the prevalence of ASD, with approximately 4% of the child population diagnosed with ASD.  

Concurrently, ADHD affects around 10% of children in Massachusetts, placing the state 23rd in the nation for the prevalence of this condition. This notable alignment in rankings underscores a matter of significant concern, warranting heightened awareness and collaborative efforts to address the challenges posed by the neuro developmental conditions within these regions. Massachusetts distinguishes itself as one of the leading states in the United States with notably low rates of bullying, ranking among the top 5 states with the least bullying. About 3.6% of children in the state experience bullying 1-2 times per month, underscoring a commendable commitment to fostering a safe and inclusive environment for the youth population.  

Population-level understanding of behavioral health is vital for public health planning and policy development. Recognizing and addressing prevalent behavioral health issues can lead to more effective public health interventions and resource allocation.